Make base controller.
